Pyarrow does not allow null buffers, even when buliding a string array with zero non-empty values.
We use std::vectors to hold the array buffer contents, and when that is empty, the data()
pointer returned is 0, causing pyarrow to barf.

Instead, pre-allocate a 0-sized buffer with the correct alignment, and if the std::vector tries to return a null pointer,
return that instead.

To improve safety (and fix a missed const), make the data() method return a const void *.